+++ backend/canisters/openchat_installer/api/src/lifecycle/init.rs
@@ -20,7 +20,6 @@ pub struct Args {
     pub registry_canister_id: CanisterId,
     pub translations_canister_id: CanisterId,
     pub website_canister_id: CanisterId,
-    pub personhood_verifier_canister_id: CanisterId,
     pub sign_in_with_email_canister_id: CanisterId,
     pub nns_governance_canister_id: CanisterId,
     pub internet_identity_canister_id: CanisterId,