+++ backend/tools/canister_installer/src/lib.rs
@@ -78,7 +78,6 @@ async fn install_service_canisters_impl(
         registry_canister_id: canister_ids.registry,
         translations_canister_id: canister_ids.translations,
         website_canister_id: canister_ids.website,
-        personhood_verifier_canister_id: canister_ids.personhood_verifier,
         sign_in_with_email_canister_id: canister_ids.sign_in_with_email,
         nns_governance_canister_id: canister_ids.nns_governance,
         internet_identity_canister_id: canister_ids.nns_internet_identity,